+++ Changelog.md
@@ -1,6 +1,14 @@
 # Vanilo Netopia Module Changelog
 
+## Unreleased
+##### 2021-XX-YY
+
+- Aligned version numbering with Vanilo core (they are in sync from now on)
+- Added mapping of payment response result to PaymentStatus
+- Added NativeStatus (returns NetopiaAction) to PaymentResponse
+
 ## 1.0.0
 ##### 2021-03-02
 
-- Initial release. Works
+- Initial release. Works.
+- Requires Vanilo 2.1